fs: Move async calls to fat_async module
