+++ src/Makefile
@@ -74,7 +74,7 @@ OS_OBJECTS = kernel/kernel_start.o kernel/syscalls.o kernel/vga.o \
 		net/ip6/ip6.o net/ip6/icmp6.o net/ip6/udp6.o net/ip6/ndp.o \
 		net/packet.o net/buffer_store.o \
 		fs/disk.o fs/filesystem.o fs/mbr.o fs/vbr.o fs/path.o \
-		fs/ext4.o fs/fat.o fs/fat_sync.o fs/memdisk.o
+		fs/ext4.o fs/fat.o fs/fat_async.o fs/fat_sync.o fs/memdisk.o
 
 CRTI_OBJ = crt/crti.o
 CRTN_OBJ = crt/crtn.o
